+++ src/endianBytes.mli
@@ -1,7 +1,7 @@
 (************************************************************************)
-(*  ocp-read                                                            *)
+(*  ocplib-endian                                                       *)
 (*                                                                      *)
-(*    Copyright 2012 OCamlPro                                           *)
+(*    Copyright 2014 OCamlPro                                           *)
 (*                                                                      *)
 (*  This file is distributed under the terms of the GNU Lesser General  *)
 (*  Public License as published by the Free Software Foundation; either *)
@@ -15,98 +15,113 @@
 (*                                                                      *)
 (************************************************************************)
 
-module type EndianStringSig = sig
+module type EndianBytesSig = sig
   (** Functions reading according to Big Endian byte order *)
 
-  val get_char : string -> int -> char
+  val get_char : Bytes.t -> int -> char
   (** [get_char buff i] reads 1 byte at offset i as a char *)
 
-  val get_uint8 : string -> int -> int
+  val get_uint8 : Bytes.t -> int -> int
   (** [get_uint8 buff i] reads 1 byte at offset i as an unsigned int of 8
   bits. i.e. It returns a value between 0 and 2^8-1 *)
 
-  val get_int8 : string -> int -> int
+  val get_int8 : Bytes.t -> int -> int
   (** [get_int8 buff i] reads 1 byte at offset i as a signed int of 8
   bits. i.e. It returns a value between -2^7 and 2^7-1 *)
 
-  val get_uint16 : string -> int -> int
+  val get_uint16 : Bytes.t -> int -> int
   (** [get_uint16 buff i] reads 2 bytes at offset i as an unsigned int
   of 16 bits. i.e. It returns a value between 0 and 2^16-1 *)
 
-  val get_int16 : string -> int -> int
+  val get_int16 : Bytes.t -> int -> int
   (** [get_int16 buff i] reads 2 byte at offset i as a signed int of
   16 bits. i.e. It returns a value between -2^15 and 2^15-1 *)
 
-  val get_int32 : string -> int -> int32
+  val get_int32 : Bytes.t -> int -> int32
   (** [get_int32 buff i] reads 4 bytes at offset i as an int32. *)
 
-  val get_int64 : string -> int -> int64
+  val get_int64 : Bytes.t -> int -> int64
   (** [get_int64 buff i] reads 8 bytes at offset i as an int64. *)
 
-  val get_float : string -> int -> float
+  val get_float : Bytes.t -> int -> float
   (** [get_float buff i] is equivalent to
       [Int32.float_of_bits (get_int32 buff i)] *)
 
-  val get_double : string -> int -> float
+  val get_double : Bytes.t -> int -> float
   (** [get_double buff i] is equivalent to
       [Int64.float_of_bits (get_int64 buff i)] *)
 
   val set_char : Bytes.t -> int -> char -> unit
-  (** @deprecated This is a deprecated alias of {!endianBytes.set_char}. *)
+  (** [set_char buff i v] writes [v] to [buff] at offset [i] *)
 
   val set_int8 : Bytes.t -> int -> int -> unit
-  (** @deprecated This is a deprecated alias of {!endianBytes.set_int8}. *)
+  (** [set_int8 buff i v] writes the least significant 8 bits of [v]
+  to [buff] at offset [i] *)
 
   val set_int16 : Bytes.t -> int -> int -> unit
-  (** @deprecated This is a deprecated alias of {!endianBytes.set_int16}. *)
+  (** [set_int16 buff i v] writes the least significant 16 bits of [v]
+  to [buff] at offset [i] *)
 
   val set_int32 : Bytes.t -> int -> int32 -> unit
-  (** @deprecated This is a deprecated alias of {!endianBytes.set_int32}. *)
+  (** [set_int32 buff i v] writes [v] to [buff] at offset [i] *)
 
   val set_int64 : Bytes.t -> int -> int64 -> unit
-  (** @deprecated This is a deprecated alias of {!endianBytes.set_int64}. *)
+  (** [set_int64 buff i v] writes [v] to [buff] at offset [i] *)
 
   val set_float : Bytes.t -> int -> float -> unit
-  (** @deprecated This is a deprecated alias of {!endianBytes.set_float}. *)
+  (** [set_float buff i v] is equivalent to
+      [set_int32 buff i (Int32.bits_of_float v)] *)
 
   val set_double : Bytes.t -> int -> float -> unit
-  (** @deprecated This is a deprecated alias of {!endianBytes.set_double}. *)
+  (** [set_double buff i v] is equivalent to
+      [set_int64 buff i (Int64.bits_of_float v)] *)
 
 end
 
-let get_char (s:string) off =
-  String.get s off
-let set_char (s:Bytes.t) off v =
-  Bytes.set s off v
-let unsafe_get_char (s:string) off =
-  String.unsafe_get s off
-let unsafe_set_char (s:Bytes.t) off v =
-  Bytes.unsafe_set s off v
+module BigEndian : sig
+  (** Functions reading according to Big Endian byte order without
+  checking for overflow *)
 
-#include "src/common.ml"
+  include EndianBytesSig
 
-#if ocaml_version >= (4, 1)
+end
 
-external unsafe_get_16 : string -> int -> int = "%caml_string_get16u"
-external unsafe_get_32 : string -> int -> int32 = "%caml_string_get32u"
-external unsafe_get_64 : string -> int -> int64 = "%caml_string_get64u"
+module BigEndian_unsafe : sig
+  (** Functions reading according to Big Endian byte order without
+  checking for overflow *)
 
-external unsafe_set_16 : Bytes.t -> int -> int -> unit = "%caml_string_set16u"
-external unsafe_set_32 : Bytes.t -> int -> int32 -> unit = "%caml_string_set32u"
-external unsafe_set_64 : Bytes.t -> int -> int64 -> unit = "%caml_string_set64u"
+  include EndianBytesSig
 
-external get_16 : string -> int -> int = "%caml_string_get16"
-external get_32 : string -> int -> int32 = "%caml_string_get32"
-external get_64 : string -> int -> int64 = "%caml_string_get64"
+end
 
-external set_16 : Bytes.t -> int -> int -> unit = "%caml_string_set16"
-external set_32 : Bytes.t -> int -> int32 -> unit = "%caml_string_set32"
-external set_64 : Bytes.t -> int -> int64 -> unit = "%caml_string_set64"
+module LittleEndian : sig
+  (** Functions reading according to Little Endian byte order *)
 
-#include "src/common_401.ml"
+  include EndianBytesSig
 
-#else
+end
 
-#include "src/common_400.ml"
+module LittleEndian_unsafe : sig
+  (** Functions reading according to Big Endian byte order without
+  checking for overflow *)
 
+  include EndianBytesSig
+
+end
+
+#if ocaml_version >= (4, 0)
+module NativeEndian : sig
+  (** Functions reading according to machine endianness *)
+
+  include EndianBytesSig
+
+end
+
+module NativeEndian_unsafe : sig
+  (** Functions reading according to machine endianness without
+  checking for overflow *)
+
+  include EndianBytesSig
+
+end
 #endif