+++ tests/bench.ml
@@ -16,80 +16,80 @@ let test_fun s f =
   Printf.printf "%s: time %f alloc: %f\n%!" s (t2 -. t1)
     (stat2.Gc.minor_words -. stat1.Gc.minor_words -. allocdiff)
 
-module String_test = struct
-  open EndianString
+module Bytes_test = struct
+  open EndianBytes
   module BE = BigEndian
   module LE = LittleEndian
 
-  let buffer = String.create buffer_size
+  let buffer = Bytes.create buffer_size
 
   let loop_read_uint16_be () =
-    for i = 0 to String.length buffer - 2 do
+    for i = 0 to Bytes.length buffer - 2 do
       ignore(BE.get_uint16 buffer i)
     done
 
   let loop_read_uint16_le () =
-    for i = 0 to String.length buffer - 2 do
+    for i = 0 to Bytes.length buffer - 2 do
       ignore(LE.get_uint16 buffer i)
     done
 
   let loop_read_int16_be () =
-    for i = 0 to String.length buffer - 2 do
+    for i = 0 to Bytes.length buffer - 2 do
       ignore(BE.get_int16 buffer i)
     done
 
   let loop_read_int16_le () =
-    for i = 0 to String.length buffer - 2 do
+    for i = 0 to Bytes.length buffer - 2 do
       ignore(LE.get_int16 buffer i)
     done
 
   let loop_read_int32_be () =
-    for i = 0 to String.length buffer - 4 do
+    for i = 0 to Bytes.length buffer - 4 do
       ignore(Int32.to_int (BE.get_int32 buffer i))
     done
 
   let loop_read_int32_le () =
-    for i = 0 to String.length buffer - 4 do
+    for i = 0 to Bytes.length buffer - 4 do
       ignore(Int32.to_int (LE.get_int32 buffer i))
     done
 
   let loop_read_int64_be () =
-    for i = 0 to String.length buffer - 8 do
+    for i = 0 to Bytes.length buffer - 8 do
       ignore(Int64.to_int (BE.get_int64 buffer i))
     done
 
   let loop_read_int64_le () =
-    for i = 0 to String.length buffer - 8 do
+    for i = 0 to Bytes.length buffer - 8 do
       ignore(Int64.to_int (LE.get_int64 buffer i))
     done
 
   let loop_write_int16_be () =
-    for i = 0 to String.length buffer - 2 do
+    for i = 0 to Bytes.length buffer - 2 do
       ignore(BE.set_int16 buffer i 10)
     done
 
   let loop_write_int16_le () =
-    for i = 0 to String.length buffer - 2 do
+    for i = 0 to Bytes.length buffer - 2 do
       ignore(LE.set_int16 buffer i 10)
     done
 
   let loop_write_int32_be () =
-    for i = 0 to String.length buffer - 4 do
+    for i = 0 to Bytes.length buffer - 4 do
       ignore((BE.set_int32 buffer i) 10l)
     done
 
   let loop_write_int32_le () =
-    for i = 0 to String.length buffer - 4 do
+    for i = 0 to Bytes.length buffer - 4 do
       ignore((LE.set_int32 buffer i) 10l)
     done
 
   let loop_write_int64_be () =
-    for i = 0 to String.length buffer - 8 do
+    for i = 0 to Bytes.length buffer - 8 do
       ignore((BE.set_int64 buffer i) 10L)
     done
 
   let loop_write_int64_le () =
-    for i = 0 to String.length buffer - 8 do
+    for i = 0 to Bytes.length buffer - 8 do
       ignore((LE.set_int64 buffer i) 10L)
     done
 
@@ -118,80 +118,80 @@ module String_test = struct
 
 end
 
-module String_unsafe_test = struct
-  open EndianString
+module Bytes_unsafe_test = struct
+  open EndianBytes
   module BE = BigEndian_unsafe
   module LE = LittleEndian_unsafe
 
-  let buffer = String.create buffer_size
+  let buffer = Bytes.create buffer_size
 
   let loop_read_uint16_be () =
-    for i = 0 to String.length buffer - 2 do
+    for i = 0 to Bytes.length buffer - 2 do
       ignore(BE.get_uint16 buffer i)
     done
 
   let loop_read_uint16_le () =
-    for i = 0 to String.length buffer - 2 do
+    for i = 0 to Bytes.length buffer - 2 do
       ignore(LE.get_uint16 buffer i)
     done
 
   let loop_read_int16_be () =
-    for i = 0 to String.length buffer - 2 do
+    for i = 0 to Bytes.length buffer - 2 do
       ignore(BE.get_int16 buffer i)
     done
 
   let loop_read_int16_le () =
-    for i = 0 to String.length buffer - 2 do
+    for i = 0 to Bytes.length buffer - 2 do
       ignore(LE.get_int16 buffer i)
     done
 
   let loop_read_int32_be () =
-    for i = 0 to String.length buffer - 4 do
+    for i = 0 to Bytes.length buffer - 4 do
       ignore(Int32.to_int (BE.get_int32 buffer i))
     done
 
   let loop_read_int32_le () =
-    for i = 0 to String.length buffer - 4 do
+    for i = 0 to Bytes.length buffer - 4 do
       ignore(Int32.to_int (LE.get_int32 buffer i))
     done
 
   let loop_read_int64_be () =
-    for i = 0 to String.length buffer - 8 do
+    for i = 0 to Bytes.length buffer - 8 do
       ignore(Int64.to_int (BE.get_int64 buffer i))
     done
 
   let loop_read_int64_le () =
-    for i = 0 to String.length buffer - 8 do
+    for i = 0 to Bytes.length buffer - 8 do
       ignore(Int64.to_int (LE.get_int64 buffer i))
     done
 
   let loop_write_int16_be () =
-    for i = 0 to String.length buffer - 2 do
+    for i = 0 to Bytes.length buffer - 2 do
       ignore(BE.set_int16 buffer i 10)
     done
 
   let loop_write_int16_le () =
-    for i = 0 to String.length buffer - 2 do
+    for i = 0 to Bytes.length buffer - 2 do
       ignore(LE.set_int16 buffer i 10)
     done
 
   let loop_write_int32_be () =
-    for i = 0 to String.length buffer - 4 do
+    for i = 0 to Bytes.length buffer - 4 do
       ignore((BE.set_int32 buffer i) 10l)
     done
 
   let loop_write_int32_le () =
-    for i = 0 to String.length buffer - 4 do
+    for i = 0 to Bytes.length buffer - 4 do
       ignore((LE.set_int32 buffer i) 10l)
     done
 
   let loop_write_int64_be () =
-    for i = 0 to String.length buffer - 8 do
+    for i = 0 to Bytes.length buffer - 8 do
       ignore((BE.set_int64 buffer i) 10L)
     done
 
   let loop_write_int64_le () =
-    for i = 0 to String.length buffer - 8 do
+    for i = 0 to Bytes.length buffer - 8 do
       ignore((LE.set_int64 buffer i) 10L)
 
    done
@@ -426,10 +426,10 @@ module Bigstring_unsafe_test = struct
 end
 
 let () =
-  Printf.printf "safe string:\n%!";
-  String_test.run_test ();
-  Printf.printf "unsafe string:\n%!";
-  String_unsafe_test.run_test ();
+  Printf.printf "safe bytes:\n%!";
+  Bytes_test.run_test ();
+  Printf.printf "unsafe bytes:\n%!";
+  Bytes_unsafe_test.run_test ();
   Printf.printf "safe bigstring:\n%!";
   Bigstring_test.run_test ();
   Printf.printf "unsafe bigstring:\n%!";